+++ golang/backend/gladiator/gladiator.go
@@ -1,6 +1,6 @@
 package gladiator
 
-import "github.com/wihrt/idle_arena/dice"
+import "github.com/wihrt/idle_arena/arena/dice"
 
 type Gladiator struct {
 	Armor                 *Armor         `json:"armor"`
